Adds option to watch specific namespaces, some refactoring
Watching resources only in some specified namespaces is frequently required in order to provide multi-tenancy or limit the reach of ingress controller for security purposes. So there's now a config option to provide a list of namespaces. In case none are provided it assumes all namespaces must be watched.

This also does some refactoring of how resources are watched, making use of API informers for ingresses and pods. It also makes sure caches get synchronised when starting the informers.

Since Kubernetes 1.14 ingresses are now part of the "networking" API instead of "extensions". While there's deprecation period, this change introduces support for both depending on the k8s version currently running.

Minor fixes and improvements added.

Fixes #27.